--- src/main/java/branch/dto/UserDto.java
@@ -1,84 +1,0 @@
-package branch.dto;
-
-public class UserDto {
-	private int num;                    
-    private String name;                
-    private String password;          
-    private String branchLocation;      
-    private String myLocation;          
-    private String phoneNum;            
-    private String grade;
-    private String profileImage;        
-    private String updatedAt;            
-    private String registratedAt;
-    private String branchNum;
-    
-    
-	public int getNum() {
-		return num;
-	}
-	public void setNum(int num) {
-		this.num = num;
-	}
-	public String getName() {
-		return name;
-	}
-	public void setName(String name) {
-		this.name = name;
-	}
-	public String getPassword() {
-		return password;
-	}
-	public void setPassword(String password) {
-		this.password = password;
-	}
-	public String getBranchLocation() {
-		return branchLocation;
-	}
-	public void setBranchLocation(String branchLocation) {
-		this.branchLocation = branchLocation;
-	}
-	public String getMyLocation() {
-		return myLocation;
-	}
-	public void setMyLocation(String myLocation) {
-		this.myLocation = myLocation;
-	}
-	public String getPhoneNum() {
-		return phoneNum;
-	}
-	public void setPhoneNum(String phoneNum) {
-		this.phoneNum = phoneNum;
-	}
-	public String getGrade() {
-		return grade;
-	}
-	public void setGrade(String grade) {
-		this.grade = grade;
-	}
-	public String getProfileImage() {
-		return profileImage;
-	}
-	public void setProfileImage(String profileImage) {
-		this.profileImage = profileImage;
-	}
-	public String getUpdatedAt() {
-		return updatedAt;
-	}
-	public void setUpdatedAt(String updatedAt) {
-		this.updatedAt = updatedAt;
-	}
-	public String getRegistratedAt() {
-		return registratedAt;
-	}
-	public void setRegistratedAt(String registratedAt) {
-		this.registratedAt = registratedAt;
-	}
-	public String getBranchNum() {
-		return branchNum;
-	}
-	public void setBranchNum(String branchNum) {
-		this.branchNum = branchNum;
-	}
-    
-}
